temp fix to DbNulls, by replacing all Nulls in DB to some int value Issue with the render function. Debug shows that all data is pulled from SQL Query and added to an array, but it will not display on the web page

+++ CommerceBank/Models/TransactionModel.cs
@@ -16,10 +16,10 @@ namespace CommerceBank.Models
         public int? AccountId { get; set; }
         [Column(TypeName = "datetime")]
         public DateTime? TransactionDate { get; set; }
-        public int? AccountBalance { get; set; }
+        public decimal? AccountBalance { get; set; }
         [StringLength(255)]
         public string TransactionType { get; set; }
-        public int? TransactionAmount { get; set; }
+        public decimal? TransactionAmount { get; set; }
         [StringLength(255)]
         public string TransactionDescription { get; set; }
         [StringLength(255)]
